Help with my problem pls (minecraft server)
I have a problem starting my server, I would greatly appreciate your help please this is the console log. Investigating there are possibilities that it is due to the ram (it has 500mb) and I try to start the server with forge, it should be noted that this problem only appears when I select the version of forge, paper or some external version like kcauldron. Thank you very much please is urgent! :(